Have only one register_view function
Refs LINK-1122
